refactor: streamline invitation retrieval and filtering logic using better-auth api
