+++ web-app/src/components/members-table/members-table.tsx
@@ -3,9 +3,10 @@
 import { useTranslations } from "next-intl";
 
 import { DataTable } from "@/components/data-table";
+import { Invitation } from "@/lib/auth/auth";
 import { InvitationStatus, MemberRole, MemberWithUser } from "@/lib/db";
 import { cn } from "@/lib/utils";
-import { Invitation, Member } from "@/prisma/generated/client";
+import { Member } from "@/prisma/generated/client";
 
 import InvitationActionsModal from "./invitation-actions-modal";
 import { InvitationActionsModalContextProvider } from "./invitation-actions-modal-context";
